// src/shiphero/generated/operations.ts
export interface PurchaseOrderQueryArgs {
  /**
   * Filter by ID; pass the public API ID unless the field name says legacy ID.
   * GraphQL type: `String`
   */

  id?: string | null;
  /**
   * Filter by PO number.
   * GraphQL type: `String`
   */

  po_number?: string | null;
  /**
   * If true, the query will only be analyzed (cost estimated) without executing it.
   * GraphQL type: `Boolean`
   */

  analyze?: boolean | null;
}

/**
 * Fetch purchase orders matching the supplied arguments.
 * Arguments for the `purchase_orders` query.
 */

export interface PurchaseOrdersQueryArgs {
  /**
   * Filter by SKU. SKU matching is case-insensitive and ignores surrounding whitespace.
   * GraphQL type: `String`
   */

  sku?: string | null;
  /**
   * Filter by public API warehouse ID. Use this ID field when selecting a warehouse; warehouse fields are display values or nested objects.
   * GraphQL type: `String`
   */

  warehouse_id?: string | null;
  /**
   * Inclusive start of the created range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  created_from?: string | null;
  /**
   * Inclusive end of the created range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  created_to?: string | null;
  /**
   * Based on updates to the Purchase Order Line Items.
   * GraphQL type: `DateTime`
   */

  updated_from?: string | null;
  /**
   * Based on updates to the Purchase Order Line Items.
   * GraphQL type: `DateTime`
   */

  updated_to?: string | null;
  /**
   * Inclusive start of the PO date range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  po_date_from?: string | null;
  /**
   * Inclusive end of the PO date range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  po_date_to?: string | null;
  /**
   * If you are a 3PL you can specify one of your customer's account.
   * GraphQL type: `String`
   */

  customer_account_id?: string | null;
  /**
   * Inclusive start of the date closed range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  date_closed_from?: string | null;
  /**
   * Inclusive end of the date closed range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  date_closed_to?: string | null;
  /**
   * Filter by fulfillment status; use values returned by the corresponding status field.
   * GraphQL type: `String`
   */

  fulfillment_status?: string | null;
  /**
   * Filter by PO number.
   * GraphQL type: `String`
   */

  po_number?: string | null;
  /**
   * If true, the query will only be analyzed (cost estimated) without executing it.
   * GraphQL type: `Boolean`
   */

  analyze?: boolean | null;
}